# speedup for gsi_obs_space