## Description
Reading gsi diag files can be pretty slow, especially for hyperspectral sounders (IASI being the worst at the moment). There are a number of things that contribute to this, but the big one is reshaping or thinning every dataset in the diag. There are other contributing factors like reading each variable for a count of nchan, making the decision to either thin, or reshape the file based on whether all values are equal in the first nchan values. 

This seems like a bad idea. It's really not hard to imagine a case where a sensor goes crazy/has fill values across all channels for an observation. If it did it on the first observation, under the current scheme, it would choose to thin rather than reshape the observation. Instead I broke the variables out into lists that are to be thinned, and reshaped ,respectively.  There is a fallback using the old scheme if the variable does not fall within the two lists. 

For a channel summary comparison between 3 runs  I have right now, under the current scheme it takes ~45 minutes, using the changes I've made here, this drops it to ~22 minutes on a dedicated node.

Timing before:
```
INFO Timers: Generate Dictionary                           0.04 seconds | Instances count: 001 | Per instance:     0.04 | Percent of total:  0.0%
INFO Timers: DataDriverExecute                          2759.65 seconds | Instances count: 003 | Per instance:   919.88 | Percent of total: 98.1%
INFO Timers: DataObjectConstructor                         0.05 seconds | Instances count: 003 | Per instance:     0.02 | Percent of total:  0.0%
INFO Timers: EvaDatasetFactory import: ...                 0.02 seconds | Instances count: 003 | Per instance:     0.01 | Percent of total:  0.0%
INFO Timers:   GsiObsSpace from ...
INFO Timers:   eva.data.gsi_obs_space
INFO Timers: DataObjectExecute                          2759.60 seconds | Instances count: 003 | Per instance:   919.87 | Percent of total: 98.1%
INFO Timers: TransformDriverExecute                       20.49 seconds | Instances count: 001 | Per instance:    20.49 | Percent of total:  0.7%
INFO Timers: Transform: accept_where                      10.95 seconds | Instances count: 006 | Per instance:     1.83 | Percent of total:  0.4%
INFO Timers: Transform: channel_stats                      9.40 seconds | Instances count: 005 | Per instance:     1.88 | Percent of total:  0.3%
INFO Timers: Transform: arithmetic                         0.01 seconds | Instances count: 006 | Per instance:     0.00 | Percent of total:  0.0%
INFO Timers: FigureDriverExecute                          15.92 seconds | Instances count: 001 | Per instance:    15.92 | Percent of total:  0.6%
INFO Timers: Graphics Loop                                 0.69 seconds | Instances count: 001 | Per instance:     0.69 | Percent of total:  0.0%
INFO Timers:
INFO Timers: Total time taken 2812.55 seconds.
```

Timing after:
```
INFO Timers:
INFO Timers: Generate Dictionary                           0.06 seconds | Instances count: 001 | Per instance:     0.06 | Percent of total:  0.0%
INFO Timers: DataDriverExecute                           873.36 seconds | Instances count: 003 | Per instance:   291.12 | Percent of total: 64.9%
INFO Timers: DataObjectConstructor                         0.03 seconds | Instances count: 003 | Per instance:     0.01 | Percent of total:  0.0%
INFO Timers: EvaDatasetFactory import: ...                 0.03 seconds | Instances count: 003 | Per instance:     0.01 | Percent of total:  0.0%
INFO Timers:   GsiObsSpace from ...
INFO Timers:   eva.data.gsi_obs_space
INFO Timers: DataObjectExecute                           873.33 seconds | Instances count: 003 | Per instance:   291.11 | Percent of total: 64.9%
INFO Timers: TransformDriverExecute                      426.81 seconds | Instances count: 001 | Per instance:   426.81 | Percent of total: 31.7%
INFO Timers: Transform: accept_where                     398.35 seconds | Instances count: 003 | Per instance:   132.78 | Percent of total: 29.6%
INFO Timers: Transform: arithmetic                         8.75 seconds | Instances count: 036 | Per instance:     0.24 | Percent of total:  0.6%
INFO Timers: Transform: channel_stats                     19.44 seconds | Instances count: 005 | Per instance:     3.89 | Percent of total:  1.4%
INFO Timers: FigureDriverExecute                          13.82 seconds | Instances count: 001 | Per instance:    13.82 | Percent of total:  1.0%
INFO Timers: Graphics Loop                                 1.28 seconds | Instances count: 001 | Per instance:     1.28 | Percent of total:  0.1%
INFO Timers:
INFO Timers: Total time taken 1345.72 seconds.
INFO Timers:


```

## Dependencies
None

## Impact
Speeds up and makes gsi_obs_space more consistent/safe among sensors.